backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
stripe==7.8.0
sendgrid==6.11.0
google-generativeai==0.3.2
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
